+++ R/class_settings.R
@@ -12,12 +12,13 @@ settings_init <- function(
   storage = "main",
   retrieval = "main"
 ) {
-  growth <- all.vars(pattern, functions = TRUE, max.names = 1L) %|||% "none"
-  dimensions <- all.vars(pattern, functions = FALSE)
+  pattern <- settings_produce_pattern(pattern)
+  dimensions <- settings_produce_dimensions(pattern)
+  settings_validate_pattern(name, pattern, dimensions)
   settings_new(
     name = name,
     format = format,
-    growth = growth,
+    pattern = pattern,
     dimensions = dimensions,
     iteration = iteration,
     error = error,
@@ -34,7 +35,7 @@ settings_init <- function(
 settings_new <- function(
   name = NULL,
   format = NULL,
-  growth = NULL,
+  pattern = NULL,
   dimensions = NULL,
   iteration = NULL,
   error = NULL,
@@ -48,7 +49,7 @@ settings_new <- function(
 ) {
   force(name)
   force(format)
-  force(growth)
+  force(pattern)
   force(dimensions)
   force(iteration)
   force(error)
@@ -62,6 +63,15 @@ settings_new <- function(
   environment()
 }
 
+settings_produce_pattern <- function(pattern) {
+  pattern <- as.expression(pattern)
+  trn(is.null(pattern[[1]]), NULL, pattern)
+}
+
+settings_produce_dimensions <- function(pattern) {
+  all.vars(pattern, functions = FALSE, unique = FALSE)
+}
+
 settings_produce_store <- function(settings) {
   store_init(settings$format, settings$resources)
 }
@@ -70,7 +80,7 @@ settings_clone <- function(settings) {
   settings_new(
     name = settings$name,
     format = settings$format,
-    growth = settings$growth,
+    pattern = settings$pattern,
     dimensions = settings$dimensions,
     iteration = settings$iteration,
     error = settings$error,
@@ -84,15 +94,20 @@ settings_clone <- function(settings) {
   )
 }
 
-settings_validate_pattern <- function(growth, dimensions) {
-  assert_scalar(growth)
-  assert_chr(growth)
-  assert_chr(dimensions)
-  if (!(growth %in% c("none", "map", "cross"))) {
-    throw_validate("pattern must be one of \"none\", \"map\", or \"cross\".")
+settings_validate_pattern <- function(name, pattern, dimensions) {
+  if (is.null(pattern)) {
+    return()
   }
-  if (!identical(growth, "none") && length(dimensions) < 1L) {
-    throw_validate("pattern must accept at least one target")
+  assert_expr(pattern)
+  assert_chr(dimensions)
+  assert_nonempty(dimensions)
+  assert_not_in(name, dimensions)
+  assert_unique(dimensions, "duplicate grouping variable in pattern.")
+  vars <- all.vars(pattern, functions = TRUE, unique = TRUE)
+  vars <- setdiff(vars, c("map", "cross", dimensions))
+  if (length(vars)) {
+    string <- string_sub_expression(deparse_safe(pattern))
+    throw_validate("invalid pattern: ", string)
   }
 }
 
@@ -100,7 +115,11 @@ settings_validate <- function(settings) {
   assert_correct_fields(settings, settings_new)
   assert_name(settings$name)
   assert_format(settings$format)
-  settings_validate_pattern(settings$growth, settings$dimensions)
+  settings_validate_pattern(
+    settings$name,
+    settings$pattern,
+    settings$dimensions
+  )
   assert_chr(settings$iteration)
   assert_in(settings$error, c("stop", "continue", "save"))
   assert_in(settings$memory, c("persistent", "transient"))